tests/wpt/tests/workers/nested_worker_terminate_from_document.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test terminating a chain of nested workers by calling terminate() from the owning document</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(function() {
  const worker = new Worker("support/parent_of_nested_worker.js");
  worker.onmessage = this.step_func_done(function(e) {
    assert_equals(e.data, "Pass");
    worker.terminate();
  });
});
</script>
