third_party/WebKit/LayoutTests/external/wpt/workers/nested_worker_terminate_from_document-expected.txt

This is a testharness.js-based test.
FAIL Test terminating a chain of nested workers by calling terminate() from the owning document assert_equals: expected "Pass" but got "Fail: ReferenceError: Worker is not defined"
Harness: the test ran to completion.
